% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonConfidenceBand.R
\name{CondPoisProcess}
\alias{CondPoisProcess}
\alias{pCondPoisProcess}
\alias{qCondPoisProcess}
\title{Conditional Poisson process}
\usage{
pCondPoisProcess(k, t, T0, N)

qCondPoisProcess(p, t, T0, N)
}
\arguments{
\item{k}{quantile}

\item{t, T0}{current and total time}

\item{N}{Total number of excesses}

\item{p}{probability}
}
\description{
This is based on \href{http://www.sciencedirect.com/science/article/pii/S0022169499001675}{Lang, Ouarda, and Bobee (1999): Towards operational guidelines for over-threshold modeling} under the hood it's the binomial distribution
Confidence limits of the Poisson process conditional on the total number of excesses
}

