% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decluster.R
\name{declusterData}
\alias{declusterData}
\title{Decluster data}
\usage{
declusterData(jdays, x, sep)
}
\arguments{
\item{jdays}{integer vector of julian days}

\item{x}{numeric vector of corresponding values}

\item{sep}{numeric separation time}
}
\description{
Decluster data
}

