% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decluster.R
\name{declusterSimpleSeparation}
\alias{declusterSimpleSeparation}
\title{Declustering using a simple separation time}
\usage{
declusterSimpleSeparation(time, x, sep)
}
\arguments{
\item{time}{numeric vector, e.g. julian days}

\item{x}{numeric vector of corresponding values}

\item{sep}{numeric separation time}
}
\value{
index integer vector of peaks
}
\description{
Declustering using a simple separation time
}

