% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGpdFlex.R
\name{fitGpdFlex}
\alias{fitGpdFlex}
\title{Fits an GPD to data}
\usage{
fitGpdFlex(data, xpar, fpar, numberOfParameters, ..., start = NULL,
  interval = NULL)
}
\arguments{
\item{data}{The data which should be modeled}

\item{xpar}{The covariates used in}

\item{fpar}{the model}

\item{numberOfParameters}{Should be omitted}

\item{...}{Further arguments passed to optim (or optimize)}

\item{start}{Vector of length numberOfParamaters}

\item{interval}{Interval if only one parameter}
}
\value{
The fitted parameters
}
\description{
Fits an GPD to data
}

