% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGpdFlex.R
\name{GetGodambeInformation}
\alias{GetGodambeInformation}
\title{Computes the Godambe information matrix}
\usage{
GetGodambeInformation(data, xpar, estimate, score)
}
\arguments{
\item{data}{Data matrix}

\item{xpar}{List Covariate information}

\item{estimate}{Vector of the parameter estimates}

\item{score}{Closure representing the score function}
}
\value{
List with \code{H} the sensitivity matrix, \code{J} the variability
matrix, and \code{G} the Godambe information
}
\description{
Computes the Godambe information matrix
}

