% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGpdFlex.R
\name{GetGodambePenalty}
\alias{GetGodambePenalty}
\title{Calculates the modified penalty using the Godambe information matrix}
\usage{
GetGodambePenalty(GodambeInformation)
}
\arguments{
\item{GodambeInformation}{e.g. the output from GetGodambeInformation}
}
\description{
Calculates the modified penalty using the Godambe information matrix
}

