% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileUncertainty.R
\name{GpdQuantileStdErr}
\alias{GpdQuantileStdErr}
\title{Computes standard error of GPD quantile (modelA)}
\usage{
GpdQuantileStdErr(period, estimates, u, lambda, cov)
}
\arguments{
\item{period}{Return period}

\item{estimates}{Model parameter estimates}

\item{u}{threshold}

\item{lambda}{Expected number of excesses per block}

\item{cov}{Covariance matrix}
}
\description{
Computes standard error of GPD quantile (modelA)
}

