% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGpdFlex.R
\name{sensitivityMatrix}
\alias{sensitivityMatrix}
\title{Calculates the sensitivity matrix}
\usage{
sensitivityMatrix(data, xpar, estimate, score, progress = TRUE)
}
\arguments{
\item{data}{Data matrix}

\item{xpar}{List Covariate information}

\item{estimate}{Vector of the parameter estimates}

\item{score}{Closure representing the score function}

\item{progress}{Should progress print to screen}
}
\description{
Calculates the sensitivity matrix
}
\note{
Close to covariance matrix (from optim) - maybe to be exhanged
}

