% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonSubsetting.R
\name{SubsetSeasons}
\alias{SubsetSeasons}
\title{Select seasonal subset and set season index}
\usage{
SubsetSeasons(dat, season)
}
\arguments{
\item{dat}{data.table of the original data}

\item{season}{String with the full season name}
}
\description{
Select seasonal subset and set season index
}

