% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGpdFlex.R
\name{variabilityMatrix}
\alias{variabilityMatrix}
\title{Calculates the variability matrix}
\usage{
variabilityMatrix(data, xpar, estimate, score, progress = TRUE)
}
\arguments{
\item{data}{Data matrix}

\item{xpar}{List Covariate information}

\item{estimate}{Vector of the parameter estimates}

\item{score}{Closure representing the score function}

\item{progress}{Should progress print to screen}
}
\description{
Calculates the variability matrix
}

